pyproject.toml
[tool.poetry]
name = "anyws"
version = "0.1.0"
description = "一个网页浏览器插件添加单词的后端"
authors = ["NSCola <3135287831@qq.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
django-environ = "^0.11.2"
drf-yasg = "^1.21.7"
django-debug-toolbar = "^4.4.6"
django-simpleui = "^2024.8.28"
djangorestframework_simplejwt = "^5.3.1"
django-solo = "2.3.0"
django-nested-inline = "^0.4.6"
django-filter = "^24.2"
django-redis = "^5.4.0"
django-reverse-admin = "^2.9.6"

#[tool.poetry.group.linux.dependencies]
#uwsgi = "^2.0.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
